@@ -1,7 +1,6 @@
 name: news_app
 description: A new Flutter project.
-authors:
-  - HASAN KOÇ
+
 
 
 publish_to: 'none'
@@ -10,7 +9,7 @@ publish_to: 'none'
 version: 1.0.0+1
 
 environment:
-  sdk: '>=2.19.6 <3.0.0'
+  sdk: '>=3.0.6 <4.0.0'
 
 
 
@@ -25,25 +24,62 @@ dependencies:
 
 
 
-  cupertino_icons: ^1.0.2
-  flutter_bloc: ^8.1.2
-  http: ^0.13.6
+  # core
   google_fonts: ^4.0.4
+  # State managment
+  flutter_bloc: ^8.1.3
+  provider: ^6.0.5
+
+  http: ^0.13.6
   url_launcher: ^6.1.10
+
+  #Localization
+  easy_localization: ^3.0.2
+
+  # 3 party uı package
   shimmer_animation: ^2.1.0+1
+
+  # json
+  json_annotation: ^4.8.1
+
+  # router
+  go_router: ^9.1.1
+  gif_view: ^0.3.1
+  google_vision: ^1.0.8+4
+  image_picker: ^1.0.1
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
+  mockito: ^5.4.2
+
+  #Build Runner
+  build_runner: ^2.4.6
+  #json
+  json_serializable: ^6.7.1
+
+  very_good_analysis: ^5.0.0+1
+
+  # uygulama iconunu generate eden eklenti
+  flutter_launcher_icons: "^0.13.1"
+
+  # basit veriler tutmak için hive kullanacagız
+  hive: ^2.2.3
+  hive_flutter: ^1.1.0
+
+  bloc_test: ^9.1.3
 
 
 
 
 
-
-  flutter_lints: ^2.0.0
-
-
+#flutter pub get
+#dart pub run flutter_launcher_icons
+flutter_launcher_icons:
+  android: "launcher_icon"
+  ios: true
+  image_path: "assets/images/news_app_icon.png"
+  min_sdk_android: 21
 
 
 
@@ -55,6 +91,11 @@ flutter:
   uses-material-design: true
 
   assets:
+    # dil json larının bulundugu yol
+    - assets/translations/
+    # uygulama içinde kullanılan assetlerin bulundugu yol
     - assets/images/
+    # uygulama içinde kullanılan giflerin bulundugu yol
+    - assets/gif/
 
 
