@@ -1,15 +1,22 @@
 import 'package:flutter/material.dart';
-import 'package:news_app/core/errors/news_error.dart';
+import 'package:news_app/bloc/news_cubit/news_error.dart';
+import 'package:news_app/core/constant/app_strings.dart';
 
-class ErrorPage extends StatelessWidget {
-  NewsError error;
-  ErrorPage({Key? key, required this.error}) : super(key: key);
+/// Hata Alındıgında gösterilecek sayfa
+@immutable
+final class ErrorPage extends StatelessWidget {
+  /// Hata Alındıgında gösterilecek sayfa
+  const ErrorPage({required this.error, super.key});
+
+  /// Hata mesajı
+  final NewsError error;
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-        body: Center(
-      child: Text("Error : ${error!.message}"),
-    ));
+      body: Center(
+        child: Text('${AppStrings.error.translate} : ${error.message}'),
+      ),
+    );
   }
 }
