@@ -1,26 +1,30 @@
 import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:news_app/bloc/news_cubit/news_cubit.dart';
+import 'package:news_app/bloc/news_cubit/news_error.dart';
+import 'package:news_app/data/interface/i_database_service.dart';
+import 'package:news_app/data/repostories/news_repositories.dart';
+import 'package:news_app/data/service/http_service.dart';
 
-import '../../core/errors/news_error.dart';
-import '../../data/interface/IDatabase_service.dart';
-import '../../data/model/news_model.dart';
-import '../../data/repostories/news_repositories.dart';
-import '../../data/service/http_service.dart';
-import 'news_cubit.dart';
-
+/// haber başlıklarının isteklerinin yapıldıgı cubit sınıfı
 class HeadlineNewsCubit extends Cubit<NewsState> {
+  /// haber başlıklarının isteklerinin yapıldıgı cubit sınıfı
   HeadlineNewsCubit() : super(NewsInitial());
 
-  NewsRepositories? _newsRepositories;
-  IDatabaseService _service = HttpService();
-  init() {
+  late NewsRepositories _newsRepositories;
+  late IDatabaseService _service;
+
+  /// kurulum methodu
+  void init() {
+    _service = HttpService();
     _newsRepositories = NewsRepositories(_service);
     getNews();
   }
 
+  /// haber başlıklarını getiren fonksiyon
   Future<void> getNews() async {
     emit(NewsLoading());
     try {
-      NewsModel data = await _newsRepositories!.getTopHeadlinesNews();
+      final data = await _newsRepositories.getTopHeadlinesNews();
       emit(NewsLoaded(data));
     } catch (e) {
       emit(NewError(NewsError(e.toString())));
