@@ -1,36 +1,48 @@
 import 'package:flutter/material.dart';
 import 'package:news_app/core/constant/app_assets.dart';
+import 'package:news_app/core/constant/app_flex.dart';
+import 'package:news_app/core/constant/app_padding.dart';
 import 'package:news_app/data/model/news_model.dart';
-import 'package:news_app/util/extension.dart';
+import 'package:news_app/ui/screens/news_detail_page.dart';
+import 'package:news_app/utils/extension.dart';
 
-import '../pages/news_detail_page.dart';
+/// yatay liste elemanı
+@immutable
+final class VerticalListItem extends StatelessWidget {
+  /// yatay liste elemanı
+  const VerticalListItem({required this.articles, super.key});
 
-class VerticalListItem extends StatelessWidget {
-  Articles articles;
-
-  VerticalListItem({Key? key, required this.articles}) : super(key: key);
+  /// haber modeli
+  final Articles articles;
 
   @override
   Widget build(BuildContext context) {
-    return SizedBox(
-        height: context.height * 0.15,
+    return Padding(
+      padding: EdgeInsets.all(AppPadding.m.padding),
+      child: SizedBox(
+        height: context.height * .15,
         child: InkWell(
           onTap: () {
-            Navigator.push(context, MaterialPageRoute(builder: (context) {
-              return NewsDetailPage(articles: articles);
-            }));
+            Navigator.push(
+              context,
+              MaterialPageRoute<void>(
+                builder: (context) {
+                  return NewsDetailPage(articles: articles);
+                },
+              ),
+            );
           },
           child: Card(
             elevation: 5,
             shape: RoundedRectangleBorder(
-              borderRadius: BorderRadius.circular(15.0),
+              borderRadius: BorderRadius.circular(15),
             ),
             clipBehavior: Clip.antiAlias,
             shadowColor: context.colorScheme.onPrimary,
             child: Row(
               children: [
                 Expanded(
-                  flex: 1,
+                  flex: AppFlex.l.flex,
                   child: articles.urlToImage == null
                       ? Image.asset(
                           AppAssets.noImage.value,
@@ -44,17 +56,16 @@ class VerticalListItem extends StatelessWidget {
                         ),
                 ),
                 Expanded(
-                  flex: 2,
+                  flex: AppFlex.l.flex,
                   child: Align(
                     alignment: Alignment.topCenter,
                     child: Padding(
-                      padding: const EdgeInsets.all(8.0),
+                      padding: EdgeInsets.all(AppPadding.m.padding),
                       child: Column(
                         mainAxisSize: MainAxisSize.min,
-                        mainAxisAlignment: MainAxisAlignment.start,
                         children: [
                           Text(
-                            articles.title ?? "",
+                            articles.title ?? '',
                             style: context.textTheme.bodyLarge,
                             softWrap: true,
                             maxLines: 2,
@@ -64,7 +75,7 @@ class VerticalListItem extends StatelessWidget {
                             height: 10,
                           ),
                           Text(
-                            articles.description ?? "",
+                            articles.description ?? '',
                             style: context.textTheme.bodySmall,
                             softWrap: true,
                             maxLines: 5,
@@ -78,6 +89,8 @@ class VerticalListItem extends StatelessWidget {
               ],
             ),
           ),
-        ));
+        ),
+      ),
+    );
   }
 }
